% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_orth.R
\name{app_orth}
\alias{app_orth}
\title{Fit Approximate Orthogonalization by Battey and Reid (2023)}
\usage{
app_orth(y, x, chosen_M, COV.S = NULL, k = 1)
}
\arguments{
\item{y}{An n-dimensional vector of outcomes.}

\item{x}{An n-dimensional vector of exposure.}

\item{chosen_M}{An n by q matrix of mediators chosen by some screening method.}

\item{COV.S}{a \code{data.frame} or \code{matrix} of covariates.}

\item{k}{A scalar for computing projection directions. Default value is 1.}
}
\value{
A a vector of test statistics (ts) each of which asymptotically follow std. normal under the null and
a vector of the corresponding p-values
}
\description{
Fit Approximate Orthogonalization by Battey and Reid (2023)
}
\examples{
data(ExampleData) # Load ExampleData1 from the package
y <- ExampleData$y # 'y' is a vector of outcomes in ExampleData
x <- ExampleData$x # 'x' is a vector of exposures in ExampleData
M <- ExampleData$M # 'M' is a matrix of mediators in ExampleData

#Perform screening using Ridge-HOLP
chosen_ind <- medsc_holp(y, x, M)
chosen_M <- M[, chosen_ind]

# Apply Approximate Orthogonalization
orth_result <- app_orth(y, x, chosen_M)
print(orth_result$ts) # Display test statistics
print(orth_result$pval) # Display p-values

}
