% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_orth_fast.R
\name{app_orth_fast}
\alias{app_orth_fast}
\title{Fast Approximate Orthogonalization}
\usage{
app_orth_fast(y, x, chosen_M, COV.S = NULL, k = 1)
}
\arguments{
\item{y}{Response vector (numeric)}

\item{x}{Main predictors (numeric matrix)}

\item{chosen_M}{Selected mediators (numeric matrix)}

\item{COV.S}{Optional covariates (numeric matrix, optional)}

\item{k}{Regularization parameter (numeric, default=1)}
}
\value{
List containing:
\itemize{
\item \code{ts}: Test statistics
\item \code{pval}: P-values
}
}
\description{
Performs approximate orthogonalization using Rcpp for improved performance.
This implements the method described by Battey and Reid (2023).
}
\examples{
data(ExampleData) # Load ExampleData1 from the package
y <- ExampleData$y # 'y' is a vector of outcomes in ExampleData
x <- ExampleData$x # 'x' is a vector of exposures in ExampleData
M <- ExampleData$M # 'M' is a matrix of mediators in ExampleData

# Perform screening using Ridge-HOLP
chosen_ind <- medsc_holp(y, x, M)
chosen_M <- M[, chosen_ind]

# Apply Approximate Orthogonalization
orth_result <- app_orth_fast(y, x, chosen_M)
print(orth_result$ts) # Display test statistics
print(orth_result$pval) # Display p-values

}
