% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_active_med.hdmt.R
\name{get_active_med.hdmt}
\alias{get_active_med.hdmt}
\title{Get Active Mediators :: Null Estimation Function From the HDMT package}
\usage{
get_active_med.hdmt(
  y,
  x,
  M,
  COV.S = NULL,
  pval.adjust = "HDMT",
  d = NULL,
  r = 1,
  k = 1
)
}
\arguments{
\item{y}{vector of outcomes}

\item{x}{vector of exposures}

\item{M}{a \code{data.frame} or \code{matrix} of mediators}

\item{COV.S}{a \code{data.frame} or \code{matrix} of covariates}

\item{pval.adjust}{specifies which method to use for controlling FWER/FDR in the joint significance testing. Either \code{'HDMT'} (default) or \code{'Bonferroni'}}

\item{d}{the number of screened mediators. Default value is \eqn{d = \cdot n/\log(n)}.}

\item{r}{a penalty parameter for the Ridge-HOLP. Default value is \code{1}}

\item{k}{a scalar for computing projection directions for AO. Default value is \code{1}.}
}
\value{
A vector of the indexes for active mediators.
}
\description{
Get Active Mediators :: Null Estimation Function From the HDMT package
}
\examples{
data(ExampleData) # Load ExampleData from the package
y <- ExampleData$y # 'y' is a vector of outcomes in ExampleData
x <- ExampleData$x #  'x' is a vector of exposures in ExampleData
M <- ExampleData$M #  'M' is a matrix of mediators in ExampleData

# Get active mediators
#using HDMT to control FDR (default)
active_mediators_index <- get_active_med.hdmt(y, x, M)

# Print the indexes of active mediators
print(active_mediators_index)


#using Bonferroni correction to control FWER
active_mediators_index <- get_active_med.hdmt(y, x, M, pval.adjust='bonferroni')

#Print the indexes of active mediators
print(active_mediators_index)
}
