% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_active_med.R
\name{get_active_med}
\alias{get_active_med}
\title{Get Active Mediators}
\usage{
get_active_med(
  y,
  x,
  M,
  COV.S = NULL,
  pval.adjust = "HDMT",
  d = NULL,
  r = 1,
  k = 1
)
}
\arguments{
\item{y}{vector of outcome}

\item{x}{vector of exposure}

\item{M}{a \code{data.frame}\\code{matrix} of mediators}

\item{COV.S}{a \code{data.frame}\\code{matrix} of covariates}

\item{pval.adjust}{specifies which method to use for controlling FWER in the joint significance testing. Either \code{'HDMT'} (default) or \code{'Bonferroni'}}

\item{d}{the number of screened mediators}

\item{r}{scalar multiplied by the identity matrix for the Ridge-HOLP. Default value is 1}

\item{k}{a scalar multiplied by identity matrix to compute projection direction for AO. Default value 1.}
}
\value{
A vector of the indexes for active mediators.
}
\description{
Get Active Mediators
}
\examples{
data(ExampleData) # Load ExampleData from the package
y <- ExampleData$y # 'y' is a vector of outcomes in ExampleData
x <- ExampleData$x # assuming 'x' is a vector of exposures in ExampleData
M <- ExampleData$M # assuming 'M' is a matrix of mediators in ExampleData

# Get active mediators
#using HDMT to control FWER (default)
active_mediators_index <- get_active_med(y, x, M)

# Print the indexes of active mediators
print(active_mediators_index)


#using Bonferroni correction to control FWER
active_mediators_index <- get_active_med(y, x, M, pval.adjust='bonferroni')

#Print the indexes of active mediators
print(active_mediators_index)
}
