% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_active_med.R
\name{get_active_med}
\alias{get_active_med}
\title{Get Active Mediators}
\usage{
get_active_med(y, x, M, COV.S = NULL, d = NULL)
}
\arguments{
\item{y}{vector of outcome}

\item{x}{vector of exposure}

\item{M}{a \code{data.frame}\\code{matrix} of mediators}

\item{COV.S}{a \code{data.frame}\\code{matrix} of covariates}

\item{d}{the number of screened mediators}
}
\value{
A vector of the indexes for active mediators.
}
\description{
Get Active Mediators
}
\examples{
data(ExampleData) # Load ExampleData from the package
y <- ExampleData$y # 'y' is a vector of outcomes in ExampleData
x <- ExampleData$x # assuming 'x' is a vector of exposures in ExampleData
M <- ExampleData$M # assuming 'M' is a matrix of mediators in ExampleData

# Get active mediators
active_index <- get_active_med(y, x, M)
print(active_index) # Print the indexes of active mediators

}
