% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medsc_holp.R
\name{medsc_holp}
\alias{medsc_holp}
\title{Screen mediators by Ridge-HOLP method}
\usage{
medsc_holp(y, x, M, COV.S = NULL, d = NULL, r = 1, approved = T)
}
\arguments{
\item{y}{n-dimensional vectors of outcome}

\item{x}{n-dimensional vectors of exposure}

\item{M}{n by p matrix of mediators}

\item{COV.S}{a \code{data.frame}\\code{matrix} of covariates}

\item{d}{desired number of chosen mediators}

\item{r}{scalar multiplied by the identity matrix for the Ridge-HOLP}

\item{approved}{temporary parameter}
}
\value{
A vector for indexes of  the selected mediators.
}
\description{
Screen mediators by Ridge-HOLP method
}
\examples{
data(ExampleData) # Load ExampleData from the data from package
y <- ExampleData$y #  'y' is a continuous vector in ExampleData
x <- ExampleData$x #   'x' is a vector in ExampleData
M <- ExampleData$M # 'M' is a matrix  of mediators in ExampleData
chosen_ind <- medsc_holp(y, x, M)
print(chosen_ind) # print indexes for the selected mediators

}
